api/requirements.txt
fastapi
uvicorn
python-jose
pyjwt
requests
cryptography